requirements.txt
psycopg2
aiogram==2.23.1
matplotlib
reportlab
chardet
pandas
Python-IO
datetime
ta
markups
logging
typing
pytz
asyncio
schedule
apimoex
apscheduler
plotly
streamlit
dash
dash_bootstrap_components
yfinance
fpdf
requests
typing
pandas_datareader
tinkoff-investments
